refactor : 체스 클래스들 패키지 수정
