+++ src/main/java/chess/game/Application.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.game;
 
 import java.util.ArrayList;
 import java.util.List;
