+++ src/main/java/chess/game/Direction.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.game;
 
 import java.util.Arrays;
 import java.util.List;
