+++ src/main/java/chess/game/Position.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.game;
 
 import java.util.Objects;
