+++ src/main/java/chess/pieces/King.java
@@ -1,10 +1,8 @@
 package chess.pieces;
 
-import static chess.Board.MAX_BOARD;
-
-import chess.Direction;
-import chess.Position;
-import chess.Rank;
+import chess.game.Direction;
+import chess.game.Position;
+import chess.game.Rank;
 import java.util.List;
 
 public class King extends Piece{
